pages refactored
