+++ src/App.css
@@ -1,0 +1,17 @@
+*{
+  margin: 0;
+  padding: 0;
+}
+
+ul{
+  margin: 0;
+  padding: 0;
+}
+
+li{
+  list-style: none;
+}
+
+body{
+  background-color: antiquewhite;
+}
