+++ src/app.js
@@ -1,11 +1,13 @@
 import React, {Fragment} from 'react';
-import Dashboard from './components/Dashboard';
-import LogIn from './components/LogIn'
+import Dashboard from './pages/Dashboard';
+import LogIn from './pages/LogIn'
 import { Routes, Route } from 'react-router-dom';
 import {AuthProvider, RequireAuth} from './contexts/AuthContext';
-import LogOut from './components/LogOut';
-import Restricted from './components/Restricted';
+import LogOut from './pages/LogOut';
+import Restricted from './pages/Restricted';
+import Home from './pages/Home';
 
+import './App.css';
 export default function App() {
   return (
   <>
@@ -15,9 +17,8 @@ export default function App() {
           <Route path='/login' element={<LogIn/>}/>
           <Route path='/logout' element={<LogOut/>}/>
           <Route path='/restricted' element={<RequireAuth><Restricted/></RequireAuth>}/>
+          <Route path='/home' element={<RequireAuth><Home/></RequireAuth>}/>
         </Routes>    
     </AuthProvider>
   </>);
 }
-
-
