+++ src/components/Sidebar.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Sidebar() {
+  return (
+    <>
+      SideNav
+    </>
+  );
+}
