+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from "react-dom";
-import App from './app';
+import App from './App';
 import { BrowserRouter } from "react-router-dom"; 
 
 ReactDOM.render(
