+++ src/pages/Home.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import Sidebar from '../components/Sidebar'
+
+export default function Home() {
+  return (
+    <div><Sidebar/> </div>
+  )
+}
